demo/requirements.txt
amazon-dax-client
aws-xray-sdk
boto3
botocore
datetime
flask
hurry.filesize
mysql-connector
redis